requirements.txt
tensorflow
numpy
pandas
fastapi
uvicorn
opencv-python
pillow
scikit-learn
matplotlib
seaborn
joblib
requests
